chore: rename NewTooltip to Tooltip
