--- packages/tradershub/src/components/NewTooltip/NewTooltip.tsx
@@ -1,23 +1,0 @@
-import React, { ReactNode } from 'react';
-import {
-    NewTooltipClassnames,
-    NewTooltipClassnamesProps,
-    NewTooltipContainerClassnames,
-    NewTooltipContainerProps,
-} from './NewTooltip.classnames';
-
-type TNewTooltip = NewTooltipClassnamesProps &
-    NewTooltipContainerProps & {
-        children: ReactNode;
-        content: ReactNode;
-    };
-
-const NewTooltip = ({ children, content, position, variant }: TNewTooltip) => (
-    <div className='relative cursor-pointer group'>
-        <div className='m-200'>{children}</div>
-        <span className={NewTooltipContainerClassnames({ position, variant })}>{content}</span>
-        <span className={NewTooltipClassnames({ position, variant })} />
-    </div>
-);
-
-export default NewTooltip;
