--- packages/tradershub/src/components/NewTooltip/index.ts
@@ -1,1 +1,0 @@
-export { default as NewTooltip } from './NewTooltip';
